+++ test/rules-methods-no_trimLeft_no_trimRight-test.js
@@ -157,8 +157,7 @@ describe('Tokenizer Rules Methods with trimLeft and trimRight disabled', functio
 
       describe('#addRule("a", 0)', function () {
         it('should trigger on empty buffer', function (done) {
-          p.addRule('a', 'consume data')
-          p.addRule(0, function (token, idx, type) {
+          p.addRule('a', 0, function (token, idx, type) {
             done()
           })
           .write('a')