+++ test/rules-methods-no_trimLeft_trimRight-test.js
@@ -156,8 +156,7 @@ describe('Tokenizer Rules Methods with trimLeft disabled', function () {
 
       describe('#addRule("a", 0)', function () {
         it('should trigger on empty buffer', function (done) {
-          p.addRule('a', 'consume data')
-          p.addRule(0, function (token, idx, type) {
+          p.addRule('a', 0, function (token, idx, type) {
             done()
           })
           .write('a')